fix: remove next-themes
