+++ package.json
@@ -23,7 +23,6 @@
     "mdx-bundler": "^9.2.1",
     "next": "13.4.19",
     "next-mdx-remote": "^4.4.1",
-    "next-themes": "^0.2.1",
     "prism-react-renderer": "^2.0.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
