+++ src/app/globals.css
@@ -5,7 +5,7 @@
 
 :root {
   --darkmode-base-color: #282930;
-  --darkmode-text-color: #cbcfdc;
+  --darkmode-text-color: #e0e0e0;
   --true-black: #000;
 }
 
@@ -30,6 +30,10 @@
   [data-theme='dark'] {
     background-color: var(--darkmode-base-color);
     color: var(--darkmode-text-color);
+  }
+
+  [data-theme='dark'] code {
+    color: var(--darkmode-base-color);
   }
 }
