+++ src/utils/set-cookie.ts
@@ -1,0 +1,8 @@
+const setCookie = (value: string) => {
+  let exdate = new Date();
+  exdate.setDate(exdate.getDate() + 7);
+  let cookieValue = value;
+  document.cookie = `theme=${cookieValue};Path=/;`;
+};
+
+export default setCookie;
